# back/api/openapi.yml
openapi: 3.0.2
info:
  title: k42un0k0blog
  version: 1.0.0
paths:
  /sign_in:
    post:
      description: Sign In
      operationId: signIn
      requestBody:
        content:
          application/json; charset=utf-8:
            schema:
              $ref: "#/components/schemas/AuthRequest"
            example:
              username: "test@test.com"
              password: "password"
      responses:
        "200":
          description: successfully authorized
          content:
            application/json; charset=utf-8:
              schema:
                $ref: "#/components/schemas/Auth"
  /auth/refresh_token:
    get:
      description: Refresh Token
      operationId: refresh token
      responses:
        "200":
          description: successfully authorized
          content:
            application/json; charset=utf-8:
              schema:
                $ref: "#/components/schemas/Auth"
        "401":
          description: authorization error
  /auth/sign_out:
    post:
      description: Sign Out
      operationId: Sign Out
      responses:
        "200":
          description: successfully authorized
          content:
            application/json; charset=utf-8:
              schema:
                type: object
                properties:
                  code: 
                    type: number
                    example: 200

        "401":
          description: authorization error
  /users:
    get:
      description: ユーザー一覧の取得
      operationId: userList
      parameters:
        - in: query
          name: page
          schema:
            type: integer
          description: 0 から始まるページ番号を指定する
      responses:
        "200":
          description: successfully request
          content:
            application/json; charset=utf-8:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserResponse"
        "401":
          description: authorization error
    post:
      description: ユーザーの作成
      operationId: userCreate
      requestBody:
        content:
          application/json; charset=utf-8:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
                password_confirmation:
                  type: string
              required:
                - name
                - email
                - password
                - password_confirmation
      responses:
        "200":
          description: successfully request
          content:
            application/json; charset=utf-8:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          description: authorization error
  /users/{id}:
    get:
      description: ユーザーの取得
      operationId: userGet
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        "200":
          description: successfully request
          content:
            application/json; charset=utf-8:
              schema:
                $ref: "#/components/schemas/UserResponse"
    put:
      description: ユーザーの更新
      operationId: userUpdate
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      requestBody:
        content:
          application/json; charset=utf-8:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
      responses:
        "200":
          description: successfully request
          content:
            application/json; charset=utf-8:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          description: authorization error
  /users/{id}/password:
    put:
      description: ユーザーのパスワードの更新
      operationId: userUpdatePassword
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      requestBody:
        content:
          application/json; charset=utf-8:
            schema:
              type: object
              properties:
                password:
                  type: string
                password_confirmation:
                  type: string
              required:
                - password
                - password_confirmation
      responses:
        "200":
          description: successfully request
        "401":
          description: authorization error
  /blogs:
    get:
      description: ブログ一覧の取得
      operationId: blogList
      responses:
        "200":
          description: successfully request
          content:
            application/json; charset=utf-8:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlogResponse"
    post:
      description: ブログの作成
      operationId: blogCreate
      requestBody:
        content:
          application/json; charset=utf-8:
            schema:
              type: object
              properties:
                title:
                  type: string
                body:
                  type: string
                blog_type:
                  type: number
              required:
                - title
                - body
                - blog_type
      responses:
        "200":
          description: successfully request
          content:
            application/json; charset=utf-8:
              schema:
                $ref: "#/components/schemas/BlogResponse"
        "401":
          description: authorization error
  /blogs/{id}:
    get:
      description: ユーザーの取得
      operationId: userGet
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        "200":
          description: successfully request
          content:
            application/json; charset=utf-8:
              schema:
                $ref: "#/components/schemas/BlogResponse"
    put:
      description: ユーザーの更新
      operationId: userUpdate
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      requestBody:
        content:
          application/json; charset=utf-8:
            schema:
              type: object
              properties:
                title:
                  type: string
                body:
                  type: string
      responses:
        "200":
          description: successfully request
          content:
            application/json; charset=utf-8:
              schema:
                $ref: "#/components/schemas/BlogResponse"
        "401":
          description: authorization error
  
components:
  schemas:
    Auth:
      type: object
      description: authorized response
      properties:
        code:
          type: number
          example: 200
        expire:
          type: string
          example: "2016-03-19T11:44:52+08:00"
        token:
          type: string
          example: "abcdefgh12345678"
      required:
        - code
        - expire
        - token
    AuthRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    UserResponse:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        email:
          type: string
      required:
        - id
        - name
        - email
    BlogResponse:
      type: object
      properties:
        id:
          type: number
        title:
          type: string
        body:
          type: string
        blog_type:
          type: number
      required:
        - id
        - title
        - body
        - blog_type